Added numba and arrow implemenatation
Now csv file uploaded is converted to arrow as an async task on the backend, and arrow file is used to read data and convert to pandas df - upto 7 times faster for a 100M row (1GB) dataset. As of now the arrow file created is 2.5 times the size of csv file. Also added switches on the front-end to toggle between numba, pandas and arrow versions. PS: numba working well only on 100M dataset, every other size below that it is outperformed by pandas. Will be updating the benchmark sheet soon