fix test failing on non-matching indices for newer dask version (#402)
When upgrading to dask 22.9.0, this test was failing as a result of a dask_dataframe compute() call resulting in a cuDF dataframe with incorrect indices (keeps partitioned indices after `map_partitions` call). 

This does not affect the behavior of graph charts in cuxfilter, so a trivial fix to the test, which compares against a result dataframe with sequential indices, is enough to resolve the issue. 

cc @galipremsagar

Authors:
  - Ajay Thorve (https://github.com/AjayThorve)

Approvers:
  - GALI PREM SAGAR (https://github.com/galipremsagar)
  - Allan (https://github.com/exactlyallan)

URL: https://github.com/rapidsai/cuxfilter/pull/402