Fix dependencies - pyproj (#514)
This PR fixes an issue raised in #511, which prevented the cuxfilter installation. This was caused by an indirect version conflict with pyproj (<3.4). 

This PR removes the `<3.4` pin, and also removes it as a direct run-dependency, since we only use it in notebooks. 

cc @exactlyallan 

resolves #511

Authors:
  - Ajay Thorve (https://github.com/AjayThorve)

Approvers:
  - AJ Schmidt (https://github.com/ajschmidt8)
  - Allan (https://github.com/exactlyallan)

URL: https://github.com/rapidsai/cuxfilter/pull/514